Move cache control into actions package
