+++ test/actions/DateTime2ToHttpDateFormatTest.scala
@@ -1,10 +1,10 @@
-package lib.httpheaders
+package actions
 
 import org.joda.time.{DateTime, DateTimeZone}
 import org.scalatest.{MustMatchers, WordSpec}
+import HttpHeaders._
 
 class DateTime2ToHttpDateFormatTest extends WordSpec with MustMatchers {
-  import lib.httpheaders._
 
   "correctly format the example date from https://www.w3.org/Protocols/rfc2616/rfc2616-sec3.html#sec3.3.1" in {
     val timestamp = new DateTime(1994, 11, 6, 8, 49, 37, DateTimeZone.UTC) // scalastyle:off magic.number
